Add UserInterface project to solution
This commit introduces a new project named "UserInterface" to the solution file `CPSC-5260-Software-Refactoring-And-Design.sln`. The project is located in `UserInterface\src\UserInterface.csproj` and is assigned the GUID `{B46E5DBB-CA75-2729-6B20-3F1457703898}`. The solution configuration platforms have also been updated to include Debug and Release build configurations for the new project on the x64 platform.