Add "Save to File" button for refactored code
Introduces a new button in the UI to allow users to save refactored code to a file when the code smell type is `DuplicatedCode`. A new method, `SaveRefactoredCodeToFile`, is added to handle the saving process, which calls the `RefactorAsync` method from the `CodeSmellDetectionService`. The UI is not yet updated with the refactored code.